glcd: Add user authentication
This is a basic, raw implementation. It does not support encryption yet
(although, the stubs are there, do not use them yet until we support TLS!).

Basically it will first check to see if the client exists. If so, it will
do the MongoDB query and then check if the password returned is equal to
the input. Pretty basic.